+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [diagram-js](https://github.com/bpmn-io/diagram-js) are d
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: emit popup menu life-cycle events
+
 ## 1.2.1
 
 * `FIX`: escape ids in CSS selectors