+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [diagram-js](https://github.com/bpmn-io/diagram-js) are d
 
 ___Note:__ Yet to be released changes appear here._
 
+## 1.3.0
+
 * `FEAT`: emit popup menu life-cycle events
 * `FIX`: prevent default click action on dragend, if `trapClick: true` is specified